tenggat waktu done

+++ application/views/teacher/penugasan.php
@@ -11,26 +11,6 @@
         <div class="col-7">
             <a href="<?= base_url(); ?>teacher/buka_tabel_nilai_tugas" class="btn btn-success mb-3">Tabel Nilai Tugas</a>
         </div>
-        <div class="col">
-        <div class="dropdown float-right">
-                <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenu2" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Tema</button>
-                <div class="dropdown-menu" aria-labelledby="dropdownMenu2">
-                    <button class="dropdown-item" type="button">Tema 1</button>
-                    <button class="dropdown-item" type="button">Tema 2</button>
-                    <button class="dropdown-item" type="button">Tema 3</button>
-                </div>
-            </div>
-        </div>
-        <div class="col">
-            <div class="dropdown float-right">
-                <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenu2" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Kelas</button>
-                <div class="dropdown-menu" aria-labelledby="dropdownMenu2">
-                    <button class="dropdown-item" type="button">Kelas 1</button>
-                    <button class="dropdown-item" type="button">Kelas 2</button>
-                    <button class="dropdown-item" type="button">Kelas 3</button>
-                </div>
-            </div>
-        </div>
     </div>
     <div class="row">
         <div class="col-lg-12">
@@ -38,6 +18,7 @@
                 <thead>
                     <tr>
                         <th scope="col">No.</th>
+                        <th scope="col">Tema</th>
                         <th scope="col">Judul Penugasan</th>
                         <th scope="col">Tenggat Pengumpulan</th>
                         <th scope="col">Opsi</th>
@@ -47,19 +28,31 @@
                     <?php $i = 1; ?>
                     <?php foreach ($penugasan as $pngs) : ?>
                         <tr>
-                            <th scope="row"><?= $i; ?></th>
-                            <td><?= $pngs['judul_penugasan']; ?></td>
-                            <td><?= date('d F Y', $pngs['date_created']); ?></td>
-                            <td>
-                                <h5>
-                                    <a href="<?= base_url(); ?>teacher/buka_daftar_tugas/<?= $pngs['id']; ?>" class="badge badge-success"> Buka </a>
-                                    <a href="<?= base_url(); ?>teacher/ubah_penugasan/<?= $pngs['id']; ?>" class="badge badge-info"> Ubah </a>
-                                    <a href="<?= base_url(); ?>teacher/hapus_penugasan/<?= $pngs['id']; ?>" data-toggle="modal" data-target="#hapus_penugasan" class="badge badge-danger"> Hapus </a>
-                                </h5>
-                            </td>
+                            <?php foreach ($kelas as $kls) : ?>
+                                <?php if ($kls['id'] == $user['kelas_id']) : ?>
+                                    <th scope="row"><?= $i; ?></th>
+                                    <td><?php foreach ($tema as $tm) : ?>
+                                            <?php
+                                            if ($tm['id'] == $pngs['tema_id']) {
+                                                echo $tm['nama_tema'];
+                                            }
+                                            ?>
+                                        <?php endforeach; ?>
+                                    </td>
+                                    <td><?= $pngs['judul_penugasan']; ?></td>
+                                    <td><?= $pngs['due_date']; ?></td>
+                                    <td>
+                                        <h5>
+                                            <a href="<?= base_url(); ?>teacher/buka_daftar_tugas/<?= $pngs['id']; ?>" class="badge badge-success"> Buka </a>
+                                            <a href="<?= base_url(); ?>teacher/ubah_penugasan/<?= $pngs['id']; ?>" class="badge badge-info"> Ubah </a>
+                                            <a href="<?= base_url(); ?>teacher/hapus_penugasan/<?= $pngs['id']; ?>" data-toggle="modal" data-target="#hapus_penugasan" class="badge badge-danger"> Hapus </a>
+                                        </h5>
+                                    </td>
                         </tr>
-                        <?php $i++; ?>
-                    <?php endforeach; ?>
+                    <?php endif; ?>
+                <?php endforeach; ?>
+                <?php $i++; ?>
+            <?php endforeach; ?>
                 </tbody>
             </table>
         </div>
